requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
cliapp==1.20180812.1
click==8.1.7
cmdtest==0.32+git
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.21.2
exceptiongroup==1.2.2
fastapi==0.114.1
Flask==3.0.3
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.2
idna==3.8
ifaddr==0.2.0
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Markdown==3.3.6
markdown2==2.5.0
MarkupSafe==2.1.5
more-itertools==8.10.0
multidict==6.1.0
netifaces==0.11.0
nicegui==2.1.0
oauthlib==3.2.0
orjson==3.10.7
pscript==0.7.7
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.4
PyYAML==5.4.1
requests==2.32.3
SecretStorage==3.3.1
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
systemd-python==234
ttystatus==0.38
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
vbuild==0.8.2
wadllib==1.3.6
waitress==3.0.0
watchfiles==0.24.0
websockets==13.0.1
Werkzeug==3.0.4
wsproto==1.2.0
yarl==1.11.1
zipp==1.0.0
